GCD of two numbers using Euclidean Algorithm
Euclidean Approach:

The idea of this algorithm is, the GCD of two numbers doesn’t change if the smaller number is subtracted from the bigger number. This is the Euclidean algorithm by subtraction. It is a process of repeat subtraction, carrying the result forward each time until the result is equal to any one number being subtracted.

Time Complexity: O(min(a,b))

Auxiliary Space: O(1)