Sieve of Eratosthenes
Sieve of Eratosthenes is the algorithm to find all the prime numbers less than or equal to a given integer n. The algorithm terminates, all the numbers in the list that are not marked are prime


Let n = 50

We create a boolean array of size n+1 from 0 to n, i.e. 0 to 50 in this case.

Sieve1

According to the algorithm we will mark all the numbers (mark them False) which are divisible by 2

sieve2

Now we move to our next unmarked number 3 and mark all the numbers which are multiples of 3

SieveofEratosthenes3

We move to our next unmarked number 5 and mark all multiples of 5

Sieve4

We continue this process and our final table will look like below:  

Sieve5

So the prime numbers are the unmarked ones: 2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31, 37, 41, 43, 47

Time Complexity: O(n*log(log(n)))


Auxiliary Space: O(n)